+++ docs/REQUIREMENTS.md
@@ -49,7 +49,7 @@
 ## 리팩터링 및 기능 수정
 
 - [x] fix: 초기 로딩 시 이름순으로 정렬
-- [ ] refactor: 컴포넌트 파일 구조 세분화
+- [x] refactor: 컴포넌트 파일 구조 세분화
 - [ ] refactor: 컴포넌트 파일명 수정
 - [ ] refactor: 이벤트 핸들러 콜백함수 분리
 - [ ] refactor: 매직넘버 상수화
