+++ src/components/restaurant/RRestaurant.ts
@@ -1,4 +1,4 @@
-import RComponent from './RComponent';
+import RComponent from '../RComponent';
 
 class RRestaurant extends RComponent {
   renderTemplate(): string {
